fix: homePage -> mainPage로 이름 수정 및 관련 모듈 수정 및 삭제
