+++ src/components/pages/mainPage.js
@@ -1,386 +1,218 @@
-const MainPAge = document.createElement("section");
+import { util } from "../../api/util";
 
-const header = document.createElement("header");
-header.classList.add("header");
+const mainPage = util.createEl("div", { class: "home-page" });
 
-const headerWrapper = document.createElement("div");
-headerWrapper.classList.add("header-wrapper");
+/**
+ * 메인 - 카테고리 메뉴
+ */
+const homeHeaderEl = util.createEl("div", { class: "home-header" });
+const homeMenusEl = util.createEl("div", { class: "home-top-menus" });
+const homeNavDivEl = util.createEl("div", { class: "home-top-div" });
+const homeNavEl = util.createEl("nav", { class: "home-top-nav" });
 
-const mainLogo = document.createElement("img");
-mainLogo.setAttribute("src", "../asset/main-logo.png");
-mainLogo.setAttribute("alt", "../main-logo");
-mainLogo.classList.add("main-logo");
+const homeEl = util.createEl("a", { class: "home-link" }, { href: "/" });
+const homeDivEl = util.createEl("div");
+const homePEl = util.createEl("p", { textContent: "홈" });
+homeDivEl.append(homePEl);
+homeEl.append(homeDivEl);
 
-const searchBtnWrapper = document.createElement("div");
-searchBtnWrapper.classList.add("search-btn-wrapper");
-
-let searchBar = document.createElement("input");
-searchBar.setAttribute("type", "text");
-searchBar.setAttribute("placeholder", "검색어를 입력해주세요.");
-searchBar.classList.add("search-bar");
-
-let inputValue = "";
-searchBar.addEventListener("input", function (e) {
-	e.preventDefault();
-	inputValue = e.target.value;
-});
-
-const searchBtn = document.createElement("img");
-searchBtn.setAttribute("src", "../asset/search-btn.png");
-searchBtn.setAttribute("alt", "search-btn");
-searchBtn.classList.add("search-btn");
-
-searchBtn.addEventListener("click", function () {
-	if (inputValue) {
-		getSitters(inputValue, 1);
-	}
-});
-
-searchBtnWrapper.appendChild(searchBar);
-searchBtnWrapper.appendChild(searchBtn);
-
-const welcomeWord = document.createElement("span");
-welcomeWord.classList.add("welcome-word");
-welcomeWord.textContent = "반가워요, 집사님!";
-
-const logout = document.createElement("span");
-logout.classList.add("logout");
-logout.textContent = "로그아웃";
-
-headerWrapper.appendChild(mainLogo);
-headerWrapper.appendChild(searchBtnWrapper);
-headerWrapper.appendChild(welcomeWord);
-headerWrapper.appendChild(logout);
-
-header.appendChild(headerWrapper);
-
-document.body.appendChild(header);
-
-const swiperContainer = document.createElement("div");
-swiperContainer.classList.add("swiper-container");
-
-const swiperWrapper = document.createElement("div");
-swiperWrapper.classList.add("swiper-wrapper");
-
-const swiperSlide1 = document.createElement("div");
-swiperSlide1.classList.add("swiper-slide");
-const img1 = document.createElement("img");
-img1.setAttribute("src", "../asset/swiper1.jpg");
-swiperSlide1.appendChild(img1);
-
-const swiperSlide2 = document.createElement("div");
-swiperSlide2.classList.add("swiper-slide");
-const img2 = document.createElement("img");
-img2.setAttribute("src", "../asset/swiper2.jpg");
-swiperSlide2.appendChild(img2);
-
-const swiperSlide3 = document.createElement("div");
-swiperSlide3.classList.add("swiper-slide");
-const img3 = document.createElement("img");
-img3.setAttribute("src", "../asset/swiper3.jpg");
-swiperSlide3.appendChild(img3);
-
-swiperWrapper.appendChild(swiperSlide1);
-swiperWrapper.appendChild(swiperSlide2);
-swiperWrapper.appendChild(swiperSlide3);
-
-const swiperButtonNext = document.createElement("div");
-swiperButtonNext.classList.add("swiper-button-next");
-
-const swiperButtonPrev = document.createElement("div");
-swiperButtonPrev.classList.add("swiper-button-prev");
-
-swiperContainer.appendChild(swiperWrapper);
-swiperContainer.appendChild(swiperButtonNext);
-swiperContainer.appendChild(swiperButtonPrev);
-
-document.body.appendChild(swiperContainer);
-
-new Swiper(".swiper-container", {
-	slidesPerView: 1,
-	loop: true,
-	navigation: {
-		nextEl: ".swiper-button-next",
-		prevEl: ".swiper-button-prev",
-	},
-});
-
-let modal = document.createElement("div");
-modal.setAttribute("id", "modal");
-const modalText = document.createElement("p");
-modalText.textContent = "로그인 하시겠습니까?";
-const modalYesBtn = document.createElement("button");
-modalYesBtn.classList.add("yes");
-modalYesBtn.textContent = "네";
-const modalNoBtn = document.createElement("button");
-modalNoBtn.classList.add("no");
-modalNoBtn.textContent = "아니오";
-
-modal.appendChild(modalText);
-modal.appendChild(modalYesBtn);
-modal.appendChild(modalNoBtn);
-
-document.body.appendChild(modal);
-
-const categories = document.createElement("div");
-categories.classList.add("categories");
-
-const hotelCategory = document.createElement("h2");
-hotelCategory.classList.add("hotel");
-hotelCategory.setAttribute(
-	"onclick",
-	"window.location.href='../pages/hotel.html'"
+const hotelEl = util.createEl(
+	"a",
+	{ class: "home-category" },
+	{ href: "/hotel" }
 );
-hotelCategory.style.color = "red";
-hotelCategory.textContent = "호텔";
+const hotelDivEl = util.createEl("div");
+const hotelPEl = util.createEl("p", { textContent: "호텔" });
+hotelDivEl.append(hotelPEl);
+hotelEl.append(hotelDivEl);
 
-const rentalCategory = document.createElement("h2");
-rentalCategory.classList.add("rental");
-rentalCategory.setAttribute(
-	"onclick",
-	"window.location.href='../pages/rental.html'"
+const rentalEl = util.createEl(
+	"a",
+	{ class: "home-category" },
+	{ href: "/rental" }
 );
-rentalCategory.textContent = "카렌트";
+const rentalDivEl = util.createEl("div");
+const rentalPEl = util.createEl("p", { textContent: "렌트카" });
+rentalDivEl.append(rentalPEl);
+rentalEl.append(rentalDivEl);
 
-const sitterCategory = document.createElement("h2");
-sitterCategory.classList.add("sitter");
-sitterCategory.setAttribute(
-	"onclick",
-	"window.location.href='../pages/sitters.html'"
+const sitterEl = util.createEl(
+	"a",
+	{ class: "home-category" },
+	{ href: "/sitter" }
 );
-sitterCategory.textContent = "펫시터";
+const sitterDivEl = util.createEl("div");
+const sitterPEl = util.createEl("p", { textContent: "펫시터" });
+sitterDivEl.append(sitterPEl);
+sitterEl.append(sitterDivEl);
 
-const snackCategory = document.createElement("h2");
-snackCategory.classList.add("snack");
-snackCategory.setAttribute(
-	"onclick",
-	"window.location.href='../pages/snack.html'"
-);
-snackCategory.textContent = "펫간식";
+const spaEl = util.createEl("a", { class: "home-category" }, { href: "/spa" });
+const spaDivEl = util.createEl("div");
+const spaPEl = util.createEl("p", { textContent: "스파" });
+spaDivEl.append(spaPEl);
+spaEl.append(spaDivEl);
 
-categories.appendChild(hotelCategory);
-categories.appendChild(rentalCategory);
-categories.appendChild(sitterCategory);
-categories.appendChild(snackCategory);
+homeNavEl.append(homeEl, hotelEl, rentalEl, sitterEl, spaEl);
+homeNavDivEl.append(homeNavEl);
+homeMenusEl.append(homeNavDivEl);
+homeHeaderEl.append(homeMenusEl);
 
-document.body.append(categories);
+/**
+ * 배너 swiper
+ */
+const swiperEl = util.createEl("div", { class: "swiper" });
+const swiperWrapper = util.createEl("div", { class: "swiper-wrapper" });
 
-const products = document.createElement("div");
-products.classList.add("products");
-
-const one = document.createElement("div");
-one.classList.add("one");
-
-products.appendChild(one);
-
-let headers = {
-	Authorization:
-		"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IkVlc3hyc0RsVW5USHkxRmpsdnZUIiwiaWF0IjoxNjc1NjkyNTI4LCJleHAiOjE2NzU3Nzg5MjgsImlzcyI6InRoZXNlY29uQGdtYWlsLmNvbSJ9.8VvD-JwUEt-YJ7LfG8P3vBZd3Zskc_1G7FJemxuJWTo",
-	"content-type": "application/json",
-	apikey: "FcKdtJs202301",
-	username: "KDT4_Team4",
-	masterKey: true,
-};
-
-//const modal = document.getElementById("modal");
-
-//const sitterEl = document.querySelector(".search-bar")
-const btn = document.querySelector(".btn");
-//const search = document.querySelector(".search-btn")
-const get = document.querySelector(".getBtn");
-
-//const res = await request("PRD08", { searchTags: ["가전"] });
-
-btn.addEventListener("click", function () {
-	addSitters();
+const swiperSlide1 = util.createEl("div", { class: "swiper-slide" });
+const swiperImg1 = util.createEl("img", {
+	src: require("../../asset/swiper1.jpg"),
 });
+swiperSlide1.append(swiperImg1);
 
-get.addEventListener("click", function () {
-	seeSitters();
+const swiperSlide2 = util.createEl("div", { class: "swiper-slide" });
+const swiperImg2 = util.createEl("img", {
+	src: require("../../asset/swiper2.jpg"),
 });
+swiperSlide2.append(swiperImg2);
 
-async function deleteSitters(todoId) {
-	const res = await fetch(
-		`https://asia-northeast3-heropy-api.cloudfunctions.net/api/products/${todoId}`,
-		{
-			method: "DELETE",
-			headers,
-		}
-	);
-	const json = await res.json();
-	console.log(json);
-}
+const swiperSlide3 = util.createEl("div", { class: "swiper-slide" });
+const swiperImg3 = util.createEl("img", {
+	src: require("../../asset/swiper3.jpg"),
+});
+swiperSlide3.append(swiperImg3);
 
-async function seeSitters() {
-	const res = await fetch(
-		"https://asia-northeast3-heropy-api.cloudfunctions.net/api/products",
-		{
-			method: "GET",
-			headers,
-		}
-	);
-	const json = await res.json();
-	console.log(json);
-}
+swiperWrapper.append(swiperSlide1);
+swiperWrapper.append(swiperSlide2);
+swiperWrapper.append(swiperSlide3);
 
-async function addSitters() {
-	for (let i = 0; i < data["list"].length; i++) {
-		const res = await fetch(
-			"https://asia-northeast3-heropy-api.cloudfunctions.net/api/products",
-			{
-				method: "POST",
-				headers,
-				body: JSON.stringify({
-					title: data["list"][i].title,
-					price: data["list"][i].price,
-					description: data["list"][i].description,
-					tags: data["list"][i].tags,
-					thumbnail: data["list"][i].thumbnailBase64,
-				}),
-			}
-		);
-		const json = await res.json();
-		console.log(json);
-	}
-}
+const swiperButtonNext = util.createEl("div", { class: "swiper-button-next" });
+const swiperButtonPrev = util.createEl("div", { class: "swiper-button-prev" });
+const swiperPagination = util.createEl("div", { class: "swiper-pagination" });
 
-// let inputValue = ''
-// sitterEl.addEventListener("input", function (e) {
-//     e.preventDefault()
-//     inputValue = e.target.value
-// })
+swiperEl.append(swiperWrapper);
+swiperEl.append(swiperButtonNext);
+swiperEl.append(swiperButtonPrev);
+swiperEl.append(swiperPagination);
 
-// search.addEventListener("click", function () {
-//     if(inputValue){
-//         getSitters(inputValue, 1)
-//     }
-// })
+/**
+ * 메인 - 상품 노출
+ * 카테고리별 4개 상품 노출(총4개 라인)
+ */
+const homeMainEl = util.createEl("div", { class: "home-main" });
+const mainLayoutEl = util.createEl("div", { class: "main-layout" });
+const mainLayoutDiv = util.createEl("div", { class: "main-layout-div" });
 
-async function getSitters(searchText, pageNumber) {
-	const res = await fetch(
-		"https://asia-northeast3-heropy-api.cloudfunctions.net/api/products/search",
-		{
-			method: "POST",
-			headers,
-			body: JSON.stringify({
-				searchText: `${searchText}`,
-				searchTags: ["호텔"],
-			}),
-		}
-	);
+// 호텔 상품 노출
+const hotelItemsEl = util.createEl("div", { class: "home-items-list" });
+// 호텔 상품 Intro
+const hotelIntroEl = util.createEl("div", { class: "home-items-intro" });
+const hotelTitleEl = util.createEl("div", { class: "intro-title" });
+const hotelTitleDiv = util.createEl("div");
+const hotelTitleStr = util.createEl("strong", {
+	textContent: "집처럼 편안한 집사와 반려묘의 쉼터 💕",
+});
+hotelTitleDiv.append(hotelTitleStr);
+hotelTitleEl.append(hotelTitleDiv);
+// 호텔 상품 더보기
+const hotelShowMoreDiv = util.createEl("div", { class: "show-more" });
+const hotelShowMoreBtn = util.createEl(
+	"button",
+	{ textContent: "더보기" },
+	{ class: "hotel-show-more" }
+);
+hotelShowMoreDiv.append(hotelShowMoreBtn);
+// 호텔 상품 라인
+const hotelItemsLineEl = util.createEl("div", { class: "home-items-line" });
+hotelItemsLineEl.classList.add("hotel-items");
+// 호텔 append
+hotelIntroEl.append(hotelTitleEl, hotelShowMoreDiv);
+hotelItemsEl.append(hotelIntroEl, hotelItemsLineEl);
 
-	const json = await res.json();
-	const result = json.map(item => item);
-	console.log(result);
+// 렌트카 상품 노출
+const rentalItemsEl = util.createEl("div", { class: "home-items-list" });
+// 렌트카 상품 Intro
+const rentalIntroEl = util.createEl("div", { class: "home-items-intro" });
+const rentalTitleEl = util.createEl("div", { class: "intro-title" });
+const rentalTitleDiv = util.createEl("div");
+const rentalTitleStr = util.createEl("strong", {
+	textContent: "🙌 부르면 언제든지 달려오는 펫택시! 🏅",
+});
+rentalTitleDiv.append(rentalTitleStr);
+rentalTitleEl.append(rentalTitleDiv);
+// 렌트카 상품 더보기
+const rentalShowMoreDiv = util.createEl("div", { class: "show-more" });
+const rentalShowMoreBtn = util.createEl(
+	"button",
+	{ textContent: "더보기" },
+	{ class: "rental-show-more" }
+);
+rentalShowMoreDiv.append(rentalShowMoreBtn);
+// 렌트카 상품 라인
+const rentalItemsLineEl = util.createEl("div", { class: "home-items-line" });
+rentalItemsLineEl.classList.add("rental-items");
+// 렌트카 append
+rentalIntroEl.append(rentalTitleEl, rentalShowMoreDiv);
+rentalItemsEl.append(rentalIntroEl, rentalItemsLineEl);
 
-	const items = [];
+// 펫시터 상품 노출
+const sitterItemsEl = util.createEl("div", { class: "home-items-list" });
+// 펫시터 상품 Intro
+const sitterIntroEl = util.createEl("div", { class: "home-items-intro" });
+const sitterTitleEl = util.createEl("div", { class: "intro-title" });
+const sitterTitleDiv = util.createEl("div");
+const sitterTitleStr = util.createEl("strong", {
+	textContent: "믿고 맡기는 전문 펫시터 서비스! 🙋‍♀️",
+});
+sitterTitleDiv.append(sitterTitleStr);
+sitterTitleEl.append(sitterTitleDiv);
+// 펫시터 상품 더보기
+const sitterShowMoreDiv = util.createEl("div", { class: "show-more" });
+const sitterShowMoreBtn = util.createEl(
+	"button",
+	{ textContent: "더보기" },
+	{ class: "sitter-show-more" }
+);
+sitterShowMoreDiv.append(sitterShowMoreBtn);
+// 펫시터 상품 라인
+const sitterItemsLineEl = util.createEl("div", { class: "home-items-line" });
+sitterItemsLineEl.classList.add("sitter-items");
+// 펫시터 append
+sitterIntroEl.append(sitterTitleEl, sitterShowMoreDiv);
+sitterItemsEl.append(sitterIntroEl, sitterItemsLineEl);
 
-	const itemsPerPage = 2;
-	const startIndex = (pageNumber - 1) * itemsPerPage;
-	const endIndex = startIndex + itemsPerPage;
+// 스파 상품 노출
+const spaItemsEl = util.createEl("div", { class: "home-items-list" });
+// 스파 상품 Intro
+const spaIntroEl = util.createEl("div", { class: "home-items-intro" });
+const spaTitleEl = util.createEl("div", { class: "intro-title" });
+const spaTitleDiv = util.createEl("div");
+const spaTitleStr = util.createEl("strong", {
+	textContent: "반려묘 전문 집사로 레벨업! 📈",
+});
+spaTitleDiv.append(spaTitleStr);
+spaTitleEl.append(spaTitleDiv);
+// 스파 상품 더보기
+const spaShowMoreDiv = util.createEl("div", { class: "show-more" });
+const spaShowMoreBtn = util.createEl(
+	"button",
+	{ textContent: "더보기" },
+	{ class: "spa-show-more" }
+);
+spaShowMoreDiv.append(spaShowMoreBtn);
+// 스파 상품 라인
+const spaItemsLineEl = util.createEl("div", { class: "home-items-line" });
+spaItemsLineEl.classList.add("spa-items");
+// 스파 append
+spaIntroEl.append(spaTitleEl, spaShowMoreDiv);
+spaItemsEl.append(spaIntroEl, spaItemsLineEl);
 
-	for (let i = startIndex; i < endIndex && i < json.length; i++) {
-		if (items.includes(json[i].title)) return;
-		if (json[i].title.includes(searchText)) {
-			items.push(json[i].title);
-			const div = document.createElement("div");
-			div.classList.add("container");
-			// div.style.height = "200px"
-			// div.style.width = "800px"
-			// div.style.margin = "20px 300px"
-			// div.style.borderTop = "1px solid lightgrey"
-			// div.style.borderBottom = "1px solid lightgrey"
+// 전체 상품 라인 append
+mainLayoutDiv.append(hotelItemsEl, rentalItemsEl, sitterItemsEl, spaItemsEl);
+mainLayoutEl.append(mainLayoutDiv);
+homeMainEl.append(mainLayoutEl);
 
-			imgEl = document.createElement("img");
-			imgEl.classList.add("mainImg");
-			imgEl.src = json[i].thumbnail;
-			// imgEl.style.width = "270px"
-			// imgEl.style.height = "120px"
-			// imgEl.style.marginLeft = "20px"
-			// imgEl.style.marginTop = "42px"
-			div.append(imgEl);
+// loading bar
+const loadingEl = util.createEl("div", { id: "loading_bar" });
 
-			imgEl.addEventListener("click", function () {
-				window.location.href = "../pages/detail.html?id=" + json[i].id;
-			});
-			// imgEl.addEventListener("click", function(e){
-			//     localStorage.setItem("variable", e.target.src);
-			//     localStorage.setItem("var2", json[i].title)
-			//     localStorage.setItem("var3", json[i].description)
-			//     window.location.href = "pages/detail.html"
-
-			// });
-
-			const title = document.createElement("h4");
-			title.classList.add("titleEl");
-			title.textContent = json[i].title;
-			// title.style.marginLeft = "300px"
-			// title.style.marginTop = "-120px"
-			div.append(title);
-
-			const desc = document.createElement("p");
-			desc.classList.add("desc");
-			// desc.style.marginLeft = "300px"
-			// desc.style.marginTop = "30px"
-			desc.textContent = json[i].description;
-			div.append(desc);
-
-			const purchaseImm = document.createElement("button");
-			purchaseImm.classList.add("baro");
-			purchaseImm.textContent = "바로구매";
-			// purchaseImm.style.marginLeft = "700px"
-			// purchaseImm.style.marginBottom = "-100px"
-			div.append(purchaseImm);
-			purchaseImm.addEventListener("click", function () {
-				modal.style.display = "block";
-				document.querySelector(".yes").addEventListener("click", function () {
-					window.location.href = "pages/login.html";
-				});
-				document.querySelector(".no").addEventListener("click", function () {
-					modal.style.display = "none";
-				});
-			});
-
-			const priceEl = document.createElement("button");
-			priceEl.classList.add("price");
-			priceEl.textContent = json[i].price + " KRW";
-			// priceEl.style.marginLeft = "640px"
-			// priceEl.style.marginTop = "20px"
-			// priceEl.style.width = "130px"
-			// priceEl.style.height = "35px"
-			// priceEl.style.backgroundColor = "green"
-			// priceEl.style.color = "white"
-			div.append(priceEl);
-			priceEl.addEventListener("click", function () {
-				window.location.href = "../pages/payment1.html";
-			});
-
-			document.body.append(div);
-		}
-	}
-	//const pagesCount = Math.ceil(json.length / itemsPerPage)
-	const paginationDiv = document.createElement("div");
-	for (let i = 1; i <= 10; i++) {
-		const buttonEl = document.createElement("button");
-		buttonEl.textContent = i;
-		if (i === pageNumber) buttonEl.disabled = true;
-		buttonEl.addEventListener("click", function () {
-			document.querySelectorAll(".container").forEach(item => item.remove());
-			getSitters(inputValue, i);
-			paginationDiv.style.position = "fixed";
-		});
-		paginationDiv.style.marginLeft = "600px";
-		paginationDiv.append(buttonEl);
-	}
-
-	document.body.append(paginationDiv);
-}
-
-// var mySwiper = new Swiper('.swiper-container', {
-//     slidesPerView: 1,
-//     loop: true,
-//     navigation: {
-//       nextEl: '.swiper-button-next',
-//       prevEl: '.swiper-button-prev',
-//     },
-//   });
+mainPage.append(homeHeaderEl, swiperEl, homeMainEl, loadingEl);
+export default mainPage;
