--- src/styles/home.css
@@ -1,3 +1,0 @@
-.home {
-  background-color: hotpink;
-}/*# sourceMappingURL=home.css.map */
