--- src/styles/home.scss
@@ -1,3 +1,0 @@
-.home {
-	background-color: hotpink;
-}
